microServices/llm-service/requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
mysqlclient==2.2.4
pymysql==1.1.0
torch>=2.1.0
transformers>=4.37.0
accelerate>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.3
chromadb>=0.4.18
sentence-transformers>=2.6.0
Pillow>=10.0.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-napoleon==0.7
django-browser-reload==1.12.1
beautifulsoup4==4.12.2